feat(utils): extended cli library to support commands and arguments (#499)
* feat(utils): extended cli library to support commands and arguments next to cli parameters
* feat(utils): improve help/usage screen for CLI
* fix(api): calling MosaicVersion.get() may return the value from previous call MosaicVersion.createFromString(), which was wrong behavior

+++ rti/mosaic-rti-core/src/main/java/org/eclipse/mosaic/rti/MosaicComponentParameters.java
@@ -35,7 +35,7 @@ public class MosaicComponentParameters {
     /**
      * The realtime break to slow down simulation for debugging/visualization purposes.
      */
-    private int realTimeBreak = 0;
+    private double realTimeBreak = 0;
 
     /**
      * Defines the number of threads to use when executing time advance of ambassadors
@@ -60,11 +60,11 @@ public class MosaicComponentParameters {
         return this;
     }
 
-    public int getRealTimeBreak() {
+    public double getRealTimeBreak() {
         return realTimeBreak;
     }
 
-    public MosaicComponentParameters setRealTimeBreak(int realTimeBreak) {
+    public MosaicComponentParameters setRealTimeBreak(double realTimeBreak) {
         this.realTimeBreak = realTimeBreak;
         return this;
     }
